% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_model.R
\docType{data}
\name{ex_model}
\alias{ex_model}
\title{Example ARIMA model}
\format{
ARIMA model
}
\source{
Arima(ex_ts,order=c(2,0,2))
}
\usage{
ex_model
}
\description{
An example AR(2) MA(2) model which can be used with backtesting functions
}
\examples{
summary(ex_model)
}
\keyword{datasets}
