% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_ts.r
\docType{data}
\name{ex_ts}
\alias{ex_ts}
\title{Example Time Series Data}
\format{
time series start from 1950:
}
\source{
arima.sim(list(order = c(2,0,2), ar = c(0.7,-0.4),ma=c(0.7,-0.4)), n = 200)
}
\usage{
ex_ts
}
\description{
Example time series data (n=200) can be used with backtesting functions
}
\keyword{datasets}
