% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getstockprice_av.R
\name{getstockprice_av}
\alias{getstockprice_av}
\title{Get Stock Price Function (AV)}
\usage{
getstockprice_av(symbol)
}
\arguments{
\item{symbol}{(character) Stock symbol in the stock market}
}
\value{
Return data frame object with historical daily stock prices
}
\description{
An API wrapper to extract historical daily price of a stock symbol from www.alphavantage.co
}
\examples{
getstockprice_av('AAPL')
}
