% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getstockprice_ms.R
\name{getstockprice_ms}
\alias{getstockprice_ms}
\title{Get Stock Price Function (MS)}
\usage{
getstockprice_ms(symbol)
}
\arguments{
\item{symbol}{(character) Stock symbol in the stock market}
}
\value{
Return data frame object with historical daily stock prices
}
\description{
An API wrapper to extract historical daily price of a stock symbol from marketstack.com
}
\examples{
getstockprice_ms('AAPL')
}
