% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_backtesting.R
\name{recursive_backtesting}
\alias{recursive_backtesting}
\title{Recursive BackTesting Function}
\usage{
recursive_backtesting(model, ts, k)
}
\arguments{
\item{model}{ARIMA model}

\item{ts}{time series data}

\item{k}{number of training data, must be fewer than total data length}
}
\value{
data as actual data
 pred as predict values
 mape as mean absolute percentage error
}
\description{
A backtesting function used for testing ARIMA model on a training time series
}
\examples{
 recursive_backtesting(ex_model,ex_ts,k=100)

}
