% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_backtesting.R
\name{window_backtesting}
\alias{window_backtesting}
\title{Moving Window BackTesting Function}
\usage{
window_backtesting(model, ts, k)
}
\arguments{
\item{model}{ARIMA model}

\item{ts}{time series data}

\item{k}{number of training data, must be fewer than total data length}
}
\value{
data as actual data
 pred as predict values
 mape as mean absolute percentage error
}
\description{
A backtesting function to test ARIMA model on a training time series
}
\examples{
 window_bactesting(ex_model,ex_ts,k=100)

}
