% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_ts.r
\docType{data}
\name{ex_ts}
\alias{ex_ts}
\title{Example Time Series Data}
\format{
A time series start from 1950
}
\source{
Generated by arima.sim(list(order = c(2,0,2), ar = c(0.7,-0.4),ma=c(0.7,-0.4)), n = 200)
}
\usage{
ex_ts
}
\description{
An example of time series data (n=200) which can be used with backtesting functions
}
\examples{
plot(ex_ts)
}
\keyword{datasets}
