% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getstockprice_ms.R
\name{getstockprice_ms}
\alias{getstockprice_ms}
\title{Get Stock Price Function (MS)}
\usage{
getstockprice_ms(symbol)
}
\arguments{
\item{symbol}{(character) stock symbol in the stock market}
}
\value{
data.frame object of stock daily historic stock price
}
\description{
an API wrapper pulling daily historic stock price data from marketstack.com
}
\examples{
getstockprice_ms('APPL')
}
